chore(api): move campaign participation result from shared to prescription context
