+++ api/src/prescription/campaign-participation/domain/models/CampaignParticipationResult.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import * as campaignParticipationService from '../../../prescription/campaign/domain/services/campaign-participation-service.js';
-import { CompetenceResult } from './CompetenceResult.js';
+import { CompetenceResult } from '../../../../shared/domain/models/CompetenceResult.js';
+import * as campaignParticipationService from '../../../campaign/domain/services/campaign-participation-service.js';
 
 class CampaignParticipationResult {
   constructor({
